Retry to load persistent stores

+++ ProtonMail/ProtonMailCommon/DataServices/CoreDataStore.swift
@@ -88,15 +88,22 @@ class CoreDataStore {
                 Analytics.shared.error(message: .coreDataError, error: ex)
                 PMLog.D(api: ex)
                 
-                do {
-                    try FileManager.default.removeItem(at: url)
-                    lastUpdatedStore.clear()
-                } catch let error as NSError{
-                    self.popError(error)
+                container.loadPersistentStores { (persistentStoreDescription, error) in
+                    if let ex = error as NSError? {
+                        Analytics.shared.error(message: .coreDataError, error: ex)
+                        PMLog.D(api: ex)
+                            
+                        do {
+                            try FileManager.default.removeItem(at: url)
+                            lastUpdatedStore.clear()
+                        } catch let error as NSError{
+                            self.popError(error)
+                        }
+                        
+                        self.popError(ex)
+                        fatalError()
+                    }
                 }
-                
-                self.popError(ex)
-                fatalError()
             } else {
                 url.excludeFromBackup()
                 container.viewContext.automaticallyMergesChangesFromParent = true
